requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.3
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
idna==3.4
Pillow==9.3.0
PyJWT==1.7.1
pytz==2022.6
requests==2.28.1
sqlparse==0.4.3
tzdata==2022.6
urllib3==1.26.13